# resources/openapi-mt.json
{
  "components": {
    "schemas": {
...
      "DataFormat": {
        "enum": [
          "JsonAoS",
          "JsonSoA",
          "JsonAoA"
        ],
        "type": "string"
      },
...
      "DatasetTailResponse": {
        "properties": {
          "data": {
            "description": "Resulting data",
            "type": "object"
          },
          "dataFormat": {
            "$ref": "#/components/schemas/DataFormat",
            "description": "How data is layed out in the response"
          },
          "schema": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Schema",
                "description": "Schema of the resulting data"
              }
            ]
          },
          "schemaFormat": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SchemaFormat",
                "description": "What representation is used for the schema"
              }
            ]
          }
        },
        "required": [
          "data",
          "dataFormat"
        ],
        "type": "object"
      },
...
      "Schema": {
        "properties": {
          "format": {
            "$ref": "#/components/schemas/SchemaFormat"
          },
          "schema": {
            "type": "object"
          }
        },
        "required": [
          "schema",
          "format"
        ],
        "type": "object"
      },
      "SchemaFormat": {
        "enum": [
          "ArrowJson",
          "Parquet",
          "ParquetJson"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "bearerFormat": "AccessToken",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "dev@kamu.dev",
      "name": "Kamu Data Inc."
    },
    "description": "Decentralized data management tool",
    "license": {
      "name": ""
    },
    "title": "kamu-cli",
    "version": "0.205.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/{account_name}/{dataset_name}/ingest": {
      "post": {
        "operationId": "dataset_ingest_handler",
        "parameters": [
          {
            "in": "path",
            "name": "sourceName",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "in": "path",
            "name": "uploadToken",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "items": {
                  "format": "int32",
                  "minimum": 0,
                  "type": "integer"
                },
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "default": null
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "api_key": []
          }
        ],
        "summary": "Push data ingestion",
        "tags": [
          "kamu"
        ]
      }
    },
...
    "/{account_name}/{dataset_name}/tail": {
      "get": {
        "operationId": "dataset_tail_handler",
        "parameters": [
          {
            "description": "Number of leading records to skip when returning result (used for\npagination)",
            "in": "path",
            "name": "skip",
            "required": true,
            "schema": {
              "format": "int64",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "description": "Maximum number of records to return (used for pagination)",
            "in": "path",
            "name": "limit",
            "required": true,
            "schema": {
              "format": "int64",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "description": "How the output data should be ecoded",
            "in": "path",
            "name": "dataFormat",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/DataFormat"
            }
          },
          {
            "description": "How to encode the schema of the result",
            "in": "path",
            "name": "schemaFormat",
            "required": true,
            "schema": {
              "oneOf": [
                {
                  "type": "null"
                },
                {
                  "$ref": "#/components/schemas/SchemaFormat"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetTailResponse"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {},
          {
            "api_key": []
          }
        ],
        "summary": "Get a sample of latest events",
        "tags": [
          "odf-query"
        ]
      }
    }
  },
...
}